+++ src/VideoPicker.php
@@ -27,7 +27,7 @@ class VideoPicker extends Plugin
     // =========================================================================
 
     public bool $hasCpSettings = true;
-    public string $schemaVersion = '1.1.0';
+    public string $schemaVersion = '1.1.1';
 
 
     // Traits